fix train
