--- backend/configs/cost_basis_10421.json
@@ -1,18 +1,0 @@
-{
-    "api_key": "4pyoUVBq2v52hdDqD6H1UI9QJwYd666qjtkPzQFW8Xgbmc0c5Q16Q00Mr4kaZdiB",
-    "api_secret": "rQeNaJhLPmEVKUurFTSvgLcuUNmOFwG1N3rN3XoF9ZcTmbq7A6O5hdjvMSmFcfv",
-    "is_testnet": true,
-    "symbols": [
-        "BTC/USDT"
-    ],
-    "timeframe": "5m",
-    "strategy": "cost_basis",
-    "strategy_params": {
-        "profit_target_pct": 5,
-        "buy_dip_pct": 10
-    },
-    "max_position_percentage": 0.2,
-    "max_trade_percentage": 0.05,
-    "check_interval": 60,
-    "maturity_level": 1
-}
