--- backend/configs/cost_basis_10421_history.json
@@ -1,1 +1,0 @@
-[]
