+++ backend/data/datasets/BTCUSDT_1h_20250516_20250517.csv
@@ -1,0 +1,26 @@
+timestamp,open,high,low,close,volume,quote_asset_volume,number_of_trades,taker_buy_base_asset_volume,taker_buy_quote_asset_volume,date
+2025-05-15 17:00:00,103796.0,104138.85,103666.4,103968.25,96.66242,10044653.3100556,4103.0,56.22143,5841686.6029057,2025-05-15
+2025-05-15 18:00:00,103968.26,103994.47,102719.88,102720.02,99.11236,10261709.1612434,3992.0,51.71294,5353702.3687025,2025-05-15
+2025-05-15 19:00:00,102719.99,103226.04,30000.0,103133.33,110.50627,11371725.5024299,4296.0,63.85845,6573986.4709499,2025-05-15
+2025-05-15 20:00:00,103133.33,103517.95,102920.58,103474.0,88.23914,9112359.3949525,3380.0,54.5879,5636327.9133714,2025-05-15
+2025-05-15 21:00:00,103474.0,103485.54,102708.46,103253.84,85.60923,8827393.9598877,3639.0,54.22264,5590107.0918557,2025-05-15
+2025-05-15 22:00:00,103189.5,123289.37,103189.42,103512.28,84.47671,8754824.764784,3719.0,59.51111,6165266.408181,2025-05-15
+2025-05-15 23:00:00,103509.01,103893.74,103248.99,103736.49,100.04826,10364596.1066444,4120.0,68.44636,7089548.6942233,2025-05-15
+2025-05-16 00:00:00,103736.49,103995.54,103522.16,103754.92,87.90142,9127538.4944017,3564.0,55.84927,5798880.0183549,2025-05-16
+2025-05-16 01:00:00,103754.83,110652.34,103471.52,104020.0,88.724,9210124.8193047,3637.0,54.07227,5612645.7075096,2025-05-16
+2025-05-16 02:00:00,104018.4,110723.74,103946.11,104198.79,101.08573,10534377.5609438,4658.0,61.55046,6413823.1421624,2025-05-16
+2025-05-16 03:00:00,104199.84,104304.85,104015.1,104023.5,72.35887,7534504.4772698,3417.0,44.35957,4618662.6412147,2025-05-16
+2025-05-16 04:00:00,104023.52,104184.65,103634.0,103911.84,73.80185,7673697.1814865,2689.0,44.09225,4584853.0291765,2025-05-16
+2025-05-16 05:00:00,103911.84,104160.35,103833.68,103916.76,77.50879,8059568.0807246,3965.0,46.00048,4783126.1477455,2025-05-16
+2025-05-16 06:00:00,103949.0,104173.55,103879.52,103987.87,74.46101,7743006.8944617,3775.0,45.68779,4750819.0378877,2025-05-16
+2025-05-16 07:00:00,103987.79,104190.47,24000.0,103710.18,72.04423,7482569.4211815,4205.0,37.43957,3891604.0882761,2025-05-16
+2025-05-16 08:00:00,103710.18,104257.2,103105.99,103410.4,103.52845,10701860.6297776,7019.0,57.41626,5934949.2907319,2025-05-16
+2025-05-16 09:00:00,103410.01,105866.7,103388.59,103800.02,102.34555,10609388.3428165,6930.0,65.33722,6772821.8771072,2025-05-16
+2025-05-16 10:00:00,103800.02,103864.01,103635.23,103802.53,106.02303,11002782.4434434,7529.0,57.99358,6018419.4303787,2025-05-16
+2025-05-16 11:00:00,103803.0,103880.37,103568.0,103641.57,99.94745,10365413.2518012,6875.0,60.34997,6259018.1802685,2025-05-16
+2025-05-16 12:00:00,103641.57,103776.71,103226.99,103506.07,103.66003,10743969.4608533,6790.0,54.89498,5689435.440982,2025-05-16
+2025-05-16 13:00:00,103506.08,104083.93,103330.64,104009.31,111.13086,11521975.1915492,7082.0,68.70358,7123549.212876,2025-05-16
+2025-05-16 14:00:00,104009.68,104010.99,103372.96,103559.36,120.95791,12541470.1023271,7850.0,71.5822,7422061.7748427,2025-05-16
+2025-05-16 15:00:00,103559.36,117750.02,103558.0,104138.11,121.13953,12596348.3690906,7645.0,75.55024,7855601.8643958,2025-05-16
+2025-05-16 16:00:00,104138.11,104541.96,103821.01,104152.68,121.68328,12671871.1684186,9267.0,68.689,7152451.5596358,2025-05-16
+2025-05-16 17:00:00,104152.68,104386.43,100845.68,103867.97,110.71658,11526801.5900085,8668.0,61.92894,6448177.2820176,2025-05-16
