--- backend/src/strategies/bollinger_bands.py
@@ -1,24 +1,0 @@
-from .base_strategy import BaseStrategy
-from typing import Dict, Any, List
-
-class Strategy(BaseStrategy):
-    def get_name(self) -> str:
-        return "Bollinger Bands"
-
-    def get_description(self) -> str:
-        return "เทรดตามการเคลื่อนไหวของราคาเมื่อเทียบกับ Bollinger Bands"
-
-    def get_parameters(self) -> List[Dict[str, Any]]:
-        return [
-            {"name": "period", "label": "Period", "type": "number", "min": 10, "max": 50, "default": 20},
-            {"name": "std_dev", "label": "Standard Deviations", "type": "number", "min": 1, "max": 3, "default": 2},
-            # เพิ่มพารามิเตอร์อื่นๆ ตามต้องการ
-        ]
-
-    def analyze(self, data: Dict[str, Any]) -> Dict[str, Any]:
-        # TODO: Implement Bollinger Bands analysis logic
-        print(f"Analyzing Bollinger Bands for {data.get('symbol')}")
-        return {
-            'action': 'HOLD',
-            'reason': 'Bollinger Bands analysis not yet implemented'
-        }
