--- backend/src/strategies/macd.py
@@ -1,25 +1,0 @@
-from .base_strategy import BaseStrategy
-from typing import Dict, Any, List
-
-class Strategy(BaseStrategy):
-    def get_name(self) -> str:
-        return "MACD"
-
-    def get_description(self) -> str:
-        return "เทรดตามสัญญาณ MACD (Moving Average Convergence Divergence)"
-
-    def get_parameters(self) -> List[Dict[str, Any]]:
-        return [
-            {"name": "fast_period", "label": "Fast EMA Period", "type": "number", "min": 5, "max": 20, "default": 12},
-            {"name": "slow_period", "label": "Slow EMA Period", "type": "number", "min": 15, "max": 50, "default": 26},
-            {"name": "signal_period", "label": "Signal EMA Period", "type": "number", "min": 5, "max": 15, "default": 9},
-            # เพิ่มพารามิเตอร์อื่นๆ ตามต้องการ
-        ]
-
-    def analyze(self, data: Dict[str, Any]) -> Dict[str, Any]:
-        # TODO: Implement MACD analysis logic
-        print(f"Analyzing MACD for {data.get('symbol')}")
-        return {
-            'action': 'HOLD',
-            'reason': 'MACD analysis not yet implemented'
-        }
