--- backend/src/strategies/rsi.py
@@ -1,25 +1,0 @@
-from .base_strategy import BaseStrategy
-from typing import Dict, Any, List
-
-class Strategy(BaseStrategy):
-    def get_name(self) -> str:
-        return "RSI"
-
-    def get_description(self) -> str:
-        return "เทรดตามสัญญาณ RSI (Relative Strength Index)"
-
-    def get_parameters(self) -> List[Dict[str, Any]]:
-        return [
-            {"name": "period", "label": "RSI Period", "type": "number", "min": 7, "max": 30, "default": 14},
-            {"name": "overbought", "label": "Overbought Threshold", "type": "number", "min": 60, "max": 90, "default": 70},
-            {"name": "oversold", "label": "Oversold Threshold", "type": "number", "min": 10, "max": 40, "default": 30},
-            # เพิ่มพารามิเตอร์อื่นๆ ตามต้องการ
-        ]
-
-    def analyze(self, data: Dict[str, Any]) -> Dict[str, Any]:
-        # TODO: Implement RSI analysis logic
-        print(f"Analyzing RSI for {data.get('symbol')}")
-        return {
-            'action': 'HOLD',
-            'reason': 'RSI analysis not yet implemented'
-        }
