--- backend/user_data/order_history.json
@@ -1,3 +1,0 @@
-{
-  "orders": []
-}
