--- backend/user_data/portfolio.json
@@ -1,5 +1,0 @@
-{
-  "last_update": "2024-03-20T10:00:00Z",
-  "total_value": 0,
-  "assets": []
-}
